Moved around some of the files to more appropiate namespaces.
