+++ src/Marvel.Api/Marvel.Api.csproj
@@ -48,36 +48,36 @@
     <Compile Include="Core.cs" />
     <Compile Include="Creators.cs" />
     <Compile Include="Events.cs" />
-    <Compile Include="Model\Character.cs" />
-    <Compile Include="Model\CharacterList.cs" />
-    <Compile Include="Model\CharacterSummary.cs" />
-    <Compile Include="Model\CollectedIssue.cs" />
-    <Compile Include="Model\Collection.cs" />
-    <Compile Include="Model\Creator.cs" />
-    <Compile Include="Model\CreatorList.cs" />
-    <Compile Include="Model\CreatorSummary.cs" />
-    <Compile Include="Model\Date.cs" />
-    <Compile Include="Model\Event.cs" />
-    <Compile Include="Model\Image.cs" />
-    <Compile Include="Model\Price.cs" />
-    <Compile Include="Model\Series.cs" />
-    <Compile Include="Model\Story.cs" />
-    <Compile Include="Model\TextObject.cs" />
-    <Compile Include="Model\Variant.cs" />
+    <Compile Include="Model\DomainObjects\Character.cs" />
+    <Compile Include="Model\Lists\CharacterList.cs" />
+    <Compile Include="Model\Summaries\CharacterSummary.cs" />
+    <Compile Include="Model\DomainObjects\CollectedIssue.cs" />
+    <Compile Include="Model\DomainObjects\Collection.cs" />
+    <Compile Include="Model\DomainObjects\Creator.cs" />
+    <Compile Include="Model\Lists\CreatorList.cs" />
+    <Compile Include="Model\Summaries\CreatorSummary.cs" />
+    <Compile Include="Model\DomainObjects\Date.cs" />
+    <Compile Include="Model\DomainObjects\Event.cs" />
+    <Compile Include="Model\DomainObjects\Image.cs" />
+    <Compile Include="Model\DomainObjects\Price.cs" />
+    <Compile Include="Model\DomainObjects\Series.cs" />
+    <Compile Include="Model\DomainObjects\Story.cs" />
+    <Compile Include="Model\DomainObjects\TextObject.cs" />
+    <Compile Include="Model\DomainObjects\Variant.cs" />
     <Compile Include="Results\CharacterResult.cs" />
-    <Compile Include="Model\ComicList.cs" />
-    <Compile Include="Model\ComicSummary.cs" />
-    <Compile Include="Model\DataContainer.cs" />
-    <Compile Include="Model\EventList.cs" />
-    <Compile Include="Model\EventSummary.cs" />
-    <Compile Include="Model\Comic .cs" />
-    <Compile Include="Model\ResultBase.cs" />
-    <Compile Include="Model\SeriesList.cs" />
-    <Compile Include="Model\SeriesSummary.cs" />
-    <Compile Include="Model\StoryList.cs" />
-    <Compile Include="Model\StorySummary.cs" />
-    <Compile Include="Model\Thumbnail.cs" />
-    <Compile Include="Model\Url.cs" />
+    <Compile Include="Model\Lists\ComicList.cs" />
+    <Compile Include="Model\Summaries\ComicSummary.cs" />
+    <Compile Include="DataContainer.cs" />
+    <Compile Include="Model\Lists\EventList.cs" />
+    <Compile Include="Model\Summaries\EventSummary.cs" />
+    <Compile Include="Model\DomainObjects\Comic .cs" />
+    <Compile Include="ResultBase.cs" />
+    <Compile Include="Model\Lists\SeriesList.cs" />
+    <Compile Include="Model\Summaries\SeriesSummary.cs" />
+    <Compile Include="Model\Lists\StoryList.cs" />
+    <Compile Include="Model\Summaries\StorySummary.cs" />
+    <Compile Include="Model\DomainObjects\Thumbnail.cs" />
+    <Compile Include="Model\DomainObjects\Url.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Results\ComicResult.cs" />
     <Compile Include="Results\CreatorResult.cs" />
