+++ src/Marvel.Api/Model/DomainObjects/CollectedIssue.cs
@@ -1,6 +1,6 @@
-﻿namespace Marvel.Api.Model
+﻿namespace Marvel.Api.Model.DomainObjects
 {
-    public class Collection
+    public class CollectedIssue
     {
         public string ResourceURI { get; set; }
         public string Name { get; set; }
