+++ src/Marvel.Api/Model/DomainObjects/Collection.cs
@@ -1,6 +1,6 @@
-﻿namespace Marvel.Api.Model
+﻿namespace Marvel.Api.Model.DomainObjects
 {
-    public class EventSummary
+    public class Collection
     {
         public string ResourceURI { get; set; }
         public string Name { get; set; }
