+++ src/Marvel.Api/Model/DomainObjects/Image.cs
@@ -1,4 +1,4 @@
-﻿namespace Marvel.Api.Model
+﻿namespace Marvel.Api.Model.DomainObjects
 {
     public class Image
     {
