+++ src/Marvel.Api/Model/DomainObjects/Price.cs
@@ -1,4 +1,4 @@
-﻿namespace Marvel.Api.Model
+﻿namespace Marvel.Api.Model.DomainObjects
 {
     public class Price
     {
