+++ src/Marvel.Api/Model/DomainObjects/TextObject.cs
@@ -1,4 +1,4 @@
-﻿namespace Marvel.Api.Model
+﻿namespace Marvel.Api.Model.DomainObjects
 {
     public class TextObject
     {
