+++ src/Marvel.Api/Model/DomainObjects/Url.cs
@@ -1,4 +1,4 @@
-﻿namespace Marvel.Api.Model
+﻿namespace Marvel.Api.Model.DomainObjects
 {
     public class Url
     {
