+++ src/Marvel.Api/Model/DomainObjects/Variant.cs
@@ -1,6 +1,6 @@
-﻿namespace Marvel.Api.Model
+﻿namespace Marvel.Api.Model.DomainObjects
 {
-    public class CollectedIssue
+    public class Variant
     {
         public string ResourceURI { get; set; }
         public string Name { get; set; }
