+++ src/Marvel.Api/Model/Lists/CharacterList.cs
@@ -1,12 +1,13 @@
 ﻿using System.Collections.Generic;
+using Marvel.Api.Model.Summaries;
 
-namespace Marvel.Api.Model
+namespace Marvel.Api.Model.Lists
 {
-    public class CreatorList
+    public class CharacterList
     {
         public string Available { get; set; }
         public string Returned { get; set; }
         public string CollectionURI { get; set; }
-        public List<CreatorSummary> Items { get; set; }    
+        public List<CharacterSummary> Items { get; set; }
     }
 }
