+++ src/Marvel.Api/Model/Lists/ComicList.cs
@@ -1,12 +1,13 @@
 ﻿using System.Collections.Generic;
+using Marvel.Api.Model.Summaries;
 
-namespace Marvel.Api.Model
+namespace Marvel.Api.Model.Lists
 {
-    public class EventList
+    public class ComicList
     {
         public string Available { get; set; }
         public string Returned { get; set; }
         public string CollectionURI { get; set; }
-        public List<EventSummary> Items { get; set; }
+        public List<ComicSummary> Items { get; set; }
     }    
 }
