+++ src/Marvel.Api/Model/Lists/EventList.cs
@@ -1,12 +1,13 @@
 ﻿using System.Collections.Generic;
+using Marvel.Api.Model.Summaries;
 
-namespace Marvel.Api.Model
+namespace Marvel.Api.Model.Lists
 {
-    public class ComicList
+    public class EventList
     {
         public string Available { get; set; }
         public string Returned { get; set; }
         public string CollectionURI { get; set; }
-        public List<ComicSummary> Items { get; set; }
+        public List<EventSummary> Items { get; set; }
     }    
 }
