+++ src/Marvel.Api/Model/Lists/SeriesList.cs
@@ -1,12 +1,13 @@
 ﻿using System.Collections.Generic;
+using Marvel.Api.Model.Summaries;
 
-namespace Marvel.Api.Model
+namespace Marvel.Api.Model.Lists
 {
-    public class StoryList
+    public class SeriesList
     {
         public string Available { get; set; }
         public string Returned { get; set; }
         public string CollectionURI { get; set; }
-        public List<StorySummary> Items { get; set; }
-    }    
+        public List<SeriesSummary> Items { get; set; }
+    }
 }
